Allowlist call between OpenStack Services
This middleware allows to allowlist network communication between OpenStack services.
If this is not configured, the middleware might rate limit requests between OpenStack services since some of them (e.g. Nova) implement an on behalf mechanism
passing on the user context. If the service starting the request runs into a rate limit, the customer will receive an HTTP Error 5xx.
E.g. Customer creating a VM in Nova, Nova calling Neutron for a port creation and runs into a neutron rate limit, does not handle the rate limit and
returns an HTTP error code 5xx to the customer instead of the rate limit error code.
To avoid this behaviour without rewriting the Openstack Services, this feature can be allowed.