Additionally set Retry-Header Header
In case of a ratelimit hit, only the `X-Retry-After` HTTP header is set.

The OpenStack terraform provider is a popular choice of many Converged Cloud
customers for building their automation on top. According to the
provider documentation the `Retry-After` header must be set to retry on
a ratelimit hit.

The `Retry-After` header is set additionally to the `X-Retry-After`
header.