Monkeypatch crc library for pyredis if necessary
crc version >= 7.0.0 renamed the attribute crc.Crc16.CCITT to
crc.Crc16.XMODEM. This attribute is explicitly used by python-redis
(import module name pyredis, but not to be confused with the pypi
package pyredis). Some users of the rate_limit middleware have pinned
their crc package to < 7, some might use a newer one. To support both
use-cases, we are now monkeypatching the crc.Crc16 enum. We do this, if
we find an existing XMODEM attribute and no CCITT attribute. In this
case we just copy over the value and everything "should be fine".

This monkeypatch can be removed, once this pr[0] has been merged and
included in a release or alternatively if we replace the python-redis
library with something that has more frequent releases.

[0] https://github.com/schlitzered/pyredis/pull/11