Don't monkeypatch crc if library is not present
We only need to monkeypatch the crc library if we have it installed.
Older versions of python-redis (pre 0.4.0) don't use crc, but crc16. To
stay compatible with these older versions where the crc is not installed
we just ignore import failures, as then there's nothing for us to do.

Needless to say, we want to either get a newer release for python-redis
or switch to a different redis library soonish.